added more data for shefflera and pilea

+++ Layer.h
@@ -3,25 +3,23 @@
 #include <vector>
 #include <cmath>
 
-
 class Layer {
 public:
-    std::vector<std::vector<double>> weights;
-    std::vector<double> biases;
-    std::vector<double> output;
-    std::vector<double> lastInput;
-    std::vector<double> deltas;
+    std::vector<std::vector<float>> weights;
+    std::vector<float> biases;
+    std::vector<float> output;
+    std::vector<float> lastInput;
+    std::vector<float> deltas;
 
     Layer(int numNeurons, int numInputsPerNeuron);
 
-    static double sigmoid(const double x) {
+    static float sigmoid(const float x) {
         return 1.0 / (1.0 + exp(-x));
     };
 
-    static double sigmoidDerivative(const double x) {
+    static float sigmoidDerivative(const float x) {
         return x * (1.0 - x);
     }
-    void forward(const std::vector<double>& inputs);
+    void forward(const std::vector<float>& inputs);
 };
-
 #endif //MLP_LAYER_H
